+++ BaiTapLonWinForm/Repositories/Interfaces/IUserRepository.cs
@@ -21,7 +21,7 @@ namespace BaiTapLonWinForm.Repositories.Interfaces
         Task<bool> SoftDeleteAsync(long userId);
         Task<bool> ExistsAsync(long userId);
         Task<bool> EmailExistsAsync(string email, long? excludeUserId = null);
-        Task<bool> ChangePassword(int userId, string newPasswordHash);
+        Task<bool> ChangePassword(long userId, string newPasswordHash);
         #endregion
 
         #region feature/ha branch
